Extract different components of LinearCrossover operator to allow for unit testing.
